+++ src/eslint.js
@@ -28,11 +28,7 @@ module.exports = {
   register(gulp, taskName, parameters) {
     gulp.task(taskName, () => {
       return gulp.src(parameters.files)
-        .pipe(eslint({
-          configFile: parameters.configFile,
-          envs: parameters.envs || {},
-          globals: parameters.globals || {}
-        }))
+        .pipe(eslint(parameters))
         .pipe(eslint.format())
         .pipe(eslint.failOnError());
     });