+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+- Fix jest coverage report
+- Enhance configuration support (eslint and jscs tasks)
+
 ## 0.1.10 (2016-01-18)
 - Introduce globals (common configurations used by multiple tasks)
 - Extended README (usage section)